+++ src/app/app.module.ts
@@ -78,6 +78,10 @@ import { MomentDatePipe } from './pipes/moment-date.pipe';
 // newly created modules
 import { HomeModule } from './home/home.module';
 import { SharedModule } from './shared/shared.module';
+import { FontAwesomeModule, FaIconLibrary } from '@fortawesome/angular-fontawesome';
+import { fas } from '@fortawesome/free-solid-svg-icons';
+import { far } from '@fortawesome/free-regular-svg-icons';
+import { fab } from '@fortawesome/free-brands-svg-icons';
 
 
 @NgModule({
@@ -156,7 +160,8 @@ import { SharedModule } from './shared/shared.module';
 		PdfViewerModule,
 		SlideshowModule,
 		HomeModule,
-		SharedModule
+		SharedModule,
+		FontAwesomeModule
 	],
 	providers: [
 		{
@@ -174,4 +179,8 @@ import { SharedModule } from './shared/shared.module';
 
 })
 export class AppModule {
+	constructor(library: FaIconLibrary) {
+		// Add an icon to the library for convenient access in other components
+		library.addIconPacks(fas, far, fab);
+	}
 }
