+++ src/app/common/day-lunch.ts
@@ -1,0 +1,10 @@
+import { MyMICDS, School, SchoolLunch } from '@mymicds/sdk';
+
+export interface DayLunch {
+	date: {
+		weekday: string,
+		date: string,
+		today: boolean
+	};
+	lunch: Record<School, SchoolLunch> | {};
+}
