+++ src/app/home/home.module.ts
@@ -3,46 +3,31 @@ import { CommonModule } from '@angular/common';
 
 import { MyMICDS, MyMICDSModule, MyMICDSModuleType } from '@mymicds/sdk';
 
+import { FontAwesomeModule, FaIconLibrary } from '@fortawesome/angular-fontawesome';
+
 import { HomeComponent } from "./home/home.component";
-
-import { moduleComponents } from './modules/module-config';
-
-import { BookmarksComponent } from './modules/bookmarks/bookmarks.component';
-import { CountdownComponent } from './modules/countdown/countdown.component';
-import { GpaCalculatorComponent } from './modules/gpa-calculator/gpa-calculator.component';
-import { ProgressComponent } from './modules/progress/progress.component';
-import { ScheduleComponent } from './modules/schedule/schedule.component';
-import { SimplifiedScheduleComponent } from './modules/simplified-schedule/simplified-schedule.component';
-import { SnowdayComponent } from './modules/snowday/snowday.component';
-import { StickynotesComponent } from './modules/stickynotes/stickynotes.component';
-import { SimplifiedLunchComponent } from './modules/simplified-lunch/simplified-lunch.component';
-import { TwitterComponent } from './modules/twitter/twitter.component';
-import { WeatherComponent } from './modules/weather/weather.component';
 
 import { SharedModule } from '../shared/shared.module';
 import { HomeRoutingModule } from './home.routing';
 
+import { fas } from '@fortawesome/free-solid-svg-icons';
+import { far } from '@fortawesome/free-regular-svg-icons';
+import { fab } from '@fortawesome/free-brands-svg-icons';
+
 @NgModule({
   declarations: [
     HomeComponent,
-    moduleComponents,
-    BookmarksComponent,
-    CountdownComponent,
-    GpaCalculatorComponent,
-    ProgressComponent,
-    ScheduleComponent,
-    SimplifiedScheduleComponent,
-    SnowdayComponent,
-    StickynotesComponent,
-    SimplifiedLunchComponent,
-    TwitterComponent,
-    WeatherComponent
   ],
   imports: [
     CommonModule,
     HomeRoutingModule,
-    SharedModule
-  ],
-  entryComponents: moduleComponents
+    SharedModule,
+    FontAwesomeModule
+  ]
 })
-export class HomeModule { }
+export class HomeModule {
+  constructor(library: FaIconLibrary) {
+    // Add an icon to the library for convenient access in other components
+    library.addIconPacks(fas, far, fab);
+  }
+}
