+++ src/app/shared/modules/bookmarks/bookmarks.component.scss
