+++ src/app/shared/modules/countdown/countdown.component.scss
