+++ src/app/shared/modules/gpa-calculator/gpa-calculator.component.scss
