+++ src/app/shared/modules/module-config.ts
