+++ src/app/shared/modules/module-options.ts
