+++ src/app/shared/modules/progress/progress.component.ts
