+++ src/app/shared/modules/schedule/schedule.component.scss
