+++ src/app/shared/modules/simplified-lunch/simplified-lunch.component.scss
