+++ src/app/shared/modules/simplified-schedule/simplified-schedule.component.html
