+++ src/app/shared/modules/snowday/snowday.component.scss
