+++ src/app/shared/modules/stickynotes/stickynotes.component.scss
