+++ src/app/shared/shared.module.ts
@@ -1,24 +1,56 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
+import { ModuleOptionsComponent } from './module-options/module-options.component';
+import { ModuleContainerComponent } from './module-container/module-container.component';
+import { ModuleOptionComponent } from './module-options/module-option/module-option.component';
+import { ModuleInspectorComponent } from './module-inspector/module-inspector.component';
+
+import { moduleComponents, config } from './modules/module-config';
+
+import { BookmarksComponent } from './modules/bookmarks/bookmarks.component';
+import { CountdownComponent } from './modules/countdown/countdown.component';
+import { GpaCalculatorComponent } from './modules/gpa-calculator/gpa-calculator.component';
+import { ProgressComponent } from './modules/progress/progress.component';
+import { ScheduleComponent } from './modules/schedule/schedule.component';
+import { SimplifiedScheduleComponent } from './modules/simplified-schedule/simplified-schedule.component';
+import { SnowdayComponent } from './modules/snowday/snowday.component';
+import { StickynotesComponent } from './modules/stickynotes/stickynotes.component';
+import { SimplifiedLunchComponent } from './modules/simplified-lunch/simplified-lunch.component';
+import { TwitterComponent } from './modules/twitter/twitter.component';
+import { WeatherComponent } from './modules/weather/weather.component';
+
 import { FontAwesomeModule, FaIconLibrary } from '@fortawesome/angular-fontawesome';
 import { fas } from '@fortawesome/free-solid-svg-icons';
 import { far } from '@fortawesome/free-regular-svg-icons';
 import { fab } from '@fortawesome/free-brands-svg-icons';
 
-import { ModuleOptionsComponent } from './module-options/module-options.component';
-import { ModuleContainerComponent } from './module-container/module-container.component';
-import { ModuleOptionComponent } from './module-options/module-option/module-option.component';
-
 @NgModule({
   declarations: [
     ModuleOptionsComponent,
     ModuleContainerComponent,
-    ModuleOptionComponent
+    ModuleOptionComponent,
+    moduleComponents,
+    BookmarksComponent,
+    CountdownComponent,
+    GpaCalculatorComponent,
+    ProgressComponent,
+    ScheduleComponent,
+    SimplifiedScheduleComponent,
+    StickynotesComponent,
+    SnowdayComponent,
+    SimplifiedLunchComponent,
+    TwitterComponent,
+    WeatherComponent,
+    ModuleInspectorComponent
   ],
   imports: [
     CommonModule,
     FontAwesomeModule
+  ],
+  entryComponents: moduleComponents,
+  exports: [
+    ModuleInspectorComponent,
   ]
 })
 export class SharedModule {
